Move DeleteConfirmationModal to its own module
